requirements.txt
google-adk>=1.0.0
asyncio>=3.4.3
typing>=3.7.4 
python-dotenv>=1.0.0
litellm>=1.0.0
pydantic>=2.0.0 